requirements.txt
# ── Core Framework ────────────────────────────────────────────────────────────
langchain>=0.2.0
langchain-community>=0.2.0

# ── Embeddings (local, open-source) ──────────────────────────────────────────
sentence-transformers>=3.0.0

# ── Vector Database (local, open-source) ─────────────────────────────────────
chromadb>=0.5.0

# ── LLM (Ollama + Qwen, local) ──────────────────────────────────────────────
# Ollama itself is installed separately: https://ollama.com
# Then run: ollama pull qwen2.5:7b

# ── Optional paid APIs (only if you want Gemini fallback) ────────────────────
google-genai>=1.0.0

# ── Web Framework ────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn>=0.30.0
pydantic>=2.7.0

# ── UI ───────────────────────────────────────────────────────────────────────
streamlit>=1.35.0

# ── Database ─────────────────────────────────────────────────────────────────
sqlalchemy>=2.0.30
bcrypt>=4.1.0

# ── PDF Processing ───────────────────────────────────────────────────────────
PyPDF2>=3.0.0
pdfplumber>=0.11.0

# ── AWS (optional) ───────────────────────────────────────────────────────────
boto3>=1.34.0

# ── Auth ─────────────────────────────────────────────────────────────────────
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9

# ── Utilities ────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
requests>=2.32.0

# ── Testing ──────────────────────────────────────────────────────────────────
pytest>=8.2.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
